backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.1.2
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.124.4
Flask==3.1.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.5.0
pydantic_core==2.27.2
PyJWT==2.8.0
pyparsing==3.3.2
python-dotenv==1.2.1
python-multipart==0.0.20
reportlab==4.2.2
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.44.0
tqdm==4.67.3
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.33.0
Werkzeug==3.1.6
